Make revisions on reports.
Mainly, changed naming conventions, moved imports, deleted unused
imports, and moved some functions between files. Also, made it so
webbrowser only pops up if --view is an option of the Entry Point.
