+++ armi/bookkeeping/report/reportingUtils.py
@@ -693,27 +693,6 @@ def makeCoreDesignReport(core, cs):
     _setGeneralSimulationData(core, cs, coreDesignTable)
 
 
-def makeCoreDesignReport2(core, cs, report):
-    r"""Builds report to summarize core design inputs
-
-    Parameters
-    ----------
-    core:  armi.reactor.reactors.Core
-    cs: armi.settings.caseSettings.Settings
-    """
-    from armi.bookkeeping import newReports
-
-    coreDesignTable = newReports.Table("Core Report Table")
-    coreDesignTable.header = ["", "Input Parameter"]
-    report["Design"]["Core Design Table"] = coreDesignTable
-
-    _setGeneralCoreDesignData2(cs, coreDesignTable)
-
-    _setGeneralCoreParametersData2(core, cs, coreDesignTable)
-
-    _setGeneralSimulationData2(core, cs, coreDesignTable)
-
-
 def _setGeneralCoreDesignData(cs, coreDesignTable):
     report.setData(
         "Case Title", "{}".format(cs.caseTitle), coreDesignTable, report.DESIGN
@@ -758,118 +737,6 @@ def _setGeneralCoreDesignData(cs, coreDesignTable):
             coreDesignTable,
             report.DESIGN,
         )
-
-
-def _setGeneralCoreDesignData2(cs, coreDesignTable):
-    coreDesignTable.addRow(["Case Title", "{}".format(cs.caseTitle)])
-    coreDesignTable.addRow(["Run Type", "{}".format(cs["runType"])])
-    coreDesignTable.addRow(["Geometry File", "{}".format(cs["geomFile"])])
-    coreDesignTable.addRow(["Loading File", "{}".format(cs["loadingFile"])])
-    coreDesignTable.addRow(
-        ["Fuel Shuffling Logic File", "{}".format(cs["shuffleLogic"])]
-    )
-    coreDesignTable.addRow(["Reactor State Loading", "{}".format(cs["loadStyle"])])
-    if cs["loadStyle"] == "fromDB":
-        coreDesignTable.addRow(["Database File", "{}".format(cs["reloadDBName"])])
-        coreDesignTable.addRow(["Starting Cycle", "{}".format(cs["startCycle"])])
-        coreDesignTable.addRow(["Starting Node", "{}".format(cs["startNode"])])
-
-
-def _setGeneralCoreParametersData2(core, cs, coreDesignTable):
-    """Sets the general Core Parameter Data
-
-    Parameters
-    ----------
-    core: Core
-    cs: Case Settings
-    coreDesignTable: newReports.Table
-        Current state of table to be added to
-    """
-    blocks = core.getBlocks()
-    totalMass = sum(b.getMass() for b in blocks)
-    fissileMass = sum(b.getFissileMass() for b in blocks)
-    heavyMetalMass = sum(b.getHMMass() for b in blocks)
-    totalVolume = sum(b.getVolume() for b in blocks)
-    coreDesignTable.addRow([" ", ""])
-    coreDesignTable.addRow(
-        ["Core Power", "{:.2f} MWth".format(cs["power"] / units.WATTS_PER_MW)]
-    )
-    coreDesignTable.addRow(
-        ["Base Capacity Factor", "{}".format(cs["availabilityFactor"])],
-    )
-    coreDesignTable.addRow(
-        ["Cycle Length", "{} days".format(cs["cycleLength"])],
-    )
-    coreDesignTable.addRow(
-        ["Burnup Cycles", "{}".format(cs["nCycles"])],
-    )
-    coreDesignTable.addRow(
-        ["Burnup Steps per Cycle", "{}".format(cs["burnSteps"])],
-    )
-    corePowerMult = int(core.powerMultiplier)
-    coreDesignTable.addRow(
-        ["Core Total Volume", "{:.2f} cc".format(totalVolume * corePowerMult)],
-    )
-    coreDesignTable.addRow(
-        [
-            "Core Fissile Mass",
-            "{:.2f} kg".format(fissileMass / units.G_PER_KG * corePowerMult),
-        ],
-    )
-    coreDesignTable.addRow(
-        [
-            "Core Heavy Metal Mass",
-            "{:.2f} kg".format(heavyMetalMass / units.G_PER_KG * corePowerMult),
-        ],
-    )
-    coreDesignTable.addRow(
-        [
-            "Core Total Mass",
-            "{:.2f} kg".format(totalMass / units.G_PER_KG * corePowerMult),
-        ]
-    )
-    coreDesignTable.addRow(
-        ["Number of Assembly Rings", "{}".format(core.getNumRings())]
-    )
-    coreDesignTable.addRow(
-        ["Number of Assemblies", "{}".format(len(core.getAssemblies() * corePowerMult))]
-    )
-    coreDesignTable.addRow(
-        [
-            "Number of Fuel Assemblies",
-            "{}".format(len(core.getAssemblies(Flags.FUEL) * corePowerMult)),
-        ]
-    )
-    coreDesignTable.addRow(
-        [
-            "Number of Control Assemblies",
-            "{}".format(len(core.getAssemblies(Flags.CONTROL) * corePowerMult)),
-        ]
-    )
-    coreDesignTable.addRow(
-        [
-            "Number of Reflector Assemblies",
-            "{}".format(len(core.getAssemblies(Flags.REFLECTOR) * corePowerMult)),
-        ]
-    )
-    coreDesignTable.addRow(
-        [
-            "Number of Shield Assemblies",
-            "{}".format(len(core.getAssemblies(Flags.SHIELD) * corePowerMult)),
-        ]
-    )
-
-
-def _setGeneralSimulationData2(core, cs, coreDesignTable):
-    coreDesignTable.addRow(["  ", ""])
-    coreDesignTable.addRow(["Full Core Model", "{}".format(core.isFullCore)])
-    coreDesignTable.addRow(
-        ["Loose Physics Coupling Enabled", "{}".format(bool(cs["looseCoupling"]))]
-    )
-    coreDesignTable.addRow(["Lattice Physics Enabled for", "{}".format(cs["genXS"])])
-    coreDesignTable.addRow(
-        ["Neutronics Enabled for", "{}".format(cs["globalFluxActive"])]
-    )
 
 
 def _setGeneralCoreParametersData(core, cs, coreDesignTable):
@@ -1027,80 +894,6 @@ def makeBlockDesignReport(r):
                     loadingFileTable,
                     report.DESIGN,
                 )
-
-
-def makeBlockDesignReport2(blueprint, report, cs):
-    from armi.reactor.components import component
-    from armi.bookkeeping import newReports
-    from armi.bookkeeping import newReportUtils
-
-    r"""Summarize the block designs from the loading file
-
-    Parameters
-    ----------
-
-    blueprint : Blueprint
-    report: ReportContent
-    cs: Case Settings
-
-    """
-    report[newReportUtils.DESIGN]["Block Summaries"] = newReports.Section(
-        "Block Summaries"
-    )
-
-    for bDesign in blueprint.blockDesigns:
-        loadingFileTable = newReports.Table(
-            "Summary Of Block: {}".format(bDesign.name), "block contents"
-        )
-        loadingFileTable.header = ["", "Input Parameter"]
-        constructedBlock = bDesign.construct(cs, blueprint, 0, 1, 0, "A", dict())
-        loadingFileTable.addRow(["Number of Components", len(bDesign)])
-        lst = [i for i in range(len(bDesign))]
-        for i, cDesign, c in zip(lst, bDesign, constructedBlock):
-            cType = cDesign.name
-            componentSplitter = (i + 1) * " " + "\n"
-            loadingFileTable.addRow([componentSplitter, ""])
-            loadingFileTable.addRow(
-                [
-                    "{} {}".format(cType, "Shape"),
-                    "{} {}".format(cDesign.shape, ""),
-                ]
-            )
-            loadingFileTable.addRow(
-                [
-                    "{} {}".format(cType, "Material"),
-                    "{} {}".format(cDesign.material, ""),
-                ]
-            )
-            loadingFileTable.addRow(
-                [
-                    "{} {}".format(cType, "Hot Temperature"),
-                    "{} {}".format(cDesign.Thot, ""),
-                ]
-            )
-            loadingFileTable.addRow(
-                [
-                    "{} {}".format(cType, "Cold Temperature"),
-                    "{} {}".format(cDesign.Tinput, ""),
-                ]
-            )
-            for pd in c.pDefs:
-                if pd.name in c.DIMENSION_NAMES:
-                    value = c.getDimension(pd.name, cold=True)
-                    unit = ""
-                    if pd.units is not None:
-                        unit = pd.units
-                    if value is not None:
-                        loadingFileTable.addRow(
-                            [
-                                "{} {}".format(cType, pd.name),
-                                "{} {}".format(value, unit),
-                            ]
-                        )
-        loadingFileTable.title = "Summary of Block: {}".format(bDesign.name)
-        report[newReportUtils.DESIGN]["Block Summaries"].addChildElement(
-            loadingFileTable, loadingFileTable.title
-        )
 
 
 def _getComponentInputDimensions(cDesign):
