--- armi/physics/neutronics/reportConstants.py
@@ -1,6 +1,0 @@
-"""
-    Constants
-"""
-NEUTRONICS_SECTION = "Neutronics"
-KEFF_PLOT = "Keff-Plot"
-INITIAL_CORE_FUEL_ASSEMBLY = "Initial Core Fuel Assembly Count"
